Revert "[7.x] Extracting dispatcher tests into files (#33203)"
This reverts commit 254c482d0c6d0bf00c7b8fcd7cf7e7c23dc40a22.
