--- tests/Events/EventsDispatcherWildcardTest.php
@@ -1,133 +1,0 @@
-<?php
-
-namespace Illuminate\Tests\Events;
-
-use Illuminate\Events\Dispatcher;
-use Mockery as m;
-use PHPUnit\Framework\TestCase;
-
-class EventsDispatcherWildcardTest extends TestCase
-{
-    protected function tearDown(): void
-    {
-        m::close();
-    }
-
-    public function testWildcardListeners()
-    {
-        unset($_SERVER['__event.test']);
-        $d = new Dispatcher;
-        $d->listen('foo.bar', function () {
-            $_SERVER['__event.test'] = 'regular';
-        });
-        $d->listen('foo.*', function () {
-            $_SERVER['__event.test'] = 'wildcard';
-        });
-        $d->listen('bar.*', function () {
-            $_SERVER['__event.test'] = 'nope';
-        });
-
-        $response = $d->dispatch('foo.bar');
-
-        $this->assertEquals([null, null], $response);
-        $this->assertSame('wildcard', $_SERVER['__event.test']);
-    }
-
-    public function testWildcardListenersWithResponses()
-    {
-        unset($_SERVER['__event.test']);
-        $d = new Dispatcher;
-        $d->listen('foo.bar', function () {
-            return 'regular';
-        });
-        $d->listen('foo.*', function () {
-            return 'wildcard';
-        });
-        $d->listen('bar.*', function () {
-            return 'nope';
-        });
-
-        $response = $d->dispatch('foo.bar');
-
-        $this->assertEquals(['regular', 'wildcard'], $response);
-    }
-
-    public function testWildcardListenersCacheFlushing()
-    {
-        unset($_SERVER['__event.test']);
-        $d = new Dispatcher;
-        $d->listen('foo.*', function () {
-            $_SERVER['__event.test'] = 'cached_wildcard';
-        });
-        $d->dispatch('foo.bar');
-        $this->assertSame('cached_wildcard', $_SERVER['__event.test']);
-
-        $d->listen('foo.*', function () {
-            $_SERVER['__event.test'] = 'new_wildcard';
-        });
-        $d->dispatch('foo.bar');
-        $this->assertSame('new_wildcard', $_SERVER['__event.test']);
-    }
-
-    public function testWildcardListenersCanBeRemoved()
-    {
-        unset($_SERVER['__event.test']);
-        $d = new Dispatcher;
-        $d->listen('foo.*', function () {
-            $_SERVER['__event.test'] = 'foo';
-        });
-        $d->forget('foo.*');
-        $d->dispatch('foo.bar');
-
-        $this->assertFalse(isset($_SERVER['__event.test']));
-    }
-
-    public function testWildcardCacheIsClearedWhenListenersAreRemoved()
-    {
-        unset($_SERVER['__event.test']);
-
-        $d = new Dispatcher;
-        $d->listen('foo*', function () {
-            $_SERVER['__event.test'] = 'foo';
-        });
-        $d->dispatch('foo');
-
-        $this->assertSame('foo', $_SERVER['__event.test']);
-
-        unset($_SERVER['__event.test']);
-
-        $d->forget('foo*');
-        $d->dispatch('foo');
-
-        $this->assertFalse(isset($_SERVER['__event.test']));
-    }
-
-    public function testWildcardListenersCanBeFound()
-    {
-        $d = new Dispatcher;
-        $this->assertFalse($d->hasListeners('foo.*'));
-
-        $d->listen('foo.*', function () {
-            //
-        });
-        $this->assertTrue($d->hasListeners('foo.*'));
-        $this->assertTrue($d->hasListeners('foo.bar'));
-    }
-
-    public function testEventPassedFirstToWildcards()
-    {
-        $d = new Dispatcher;
-        $d->listen('foo.*', function ($event, $data) {
-            $this->assertSame('foo.bar', $event);
-            $this->assertEquals(['first', 'second'], $data);
-        });
-        $d->dispatch('foo.bar', ['first', 'second']);
-
-        $d = new Dispatcher;
-        $d->listen('foo.bar', function ($first, $second) {
-            $this->assertSame('first', $first);
-            $this->assertSame('second', $second);
-        });
-        $d->dispatch('foo.bar', ['first', 'second']);
-    }
-}
